function that inserts a new node at a given position

+++ 0x13-more_singly_linked_lists/9-insert_nodeint.c
@@ -22,9 +22,8 @@ listint_t *insert_nodeint_at_index(listint_t **head, unsigned int idx, int n)
 		new = malloc(sizeof(listint_t));
 		if (!new)
 			return (NULL);
-		new->n = (int)n;
+		new->n = n;
 		new->next = *head;
-
 		*head = new;
 
 		return (new);
